backend/requirements_packaged.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
waitress==2.1.2
psycopg2-binary==2.9.7
Werkzeug==2.3.7
Jinja2==3.1.2